+++ java/shared/java/org/signal/libsignal/internal/NativeTesting.kt
@@ -387,6 +387,9 @@ public object NativeTesting {
   public external fun TESTING_GetMessageBackupInfoTests(): Array<Object>
 
   @JvmStatic
+  public external fun TESTING_GetPreKeyCountTests(): Array<Object>
+
+  @JvmStatic
   public external fun TESTING_InputStreamReadIntoZeroLengthSlice(capsAlphabetInput: InputStream): ByteArray
 
   @JvmStatic