+++ rust/net/chat/src/grpc.rs
@@ -11,6 +11,7 @@ pub mod backups;
 pub mod call_quality;
 pub mod credentials;
 pub mod devices;
+pub mod keys;
 mod messages;
 pub mod payments;
 mod profiles;